feat(frontend): server-prefetch remaining list pages
Tools, Models, Projects, Tokens, Architect, Explorer, Traces and the
Organization Settings team tab now fetch their first page in an async
server component and hand it to a client component as initialData, so
the first paint shows content (or the real empty state) instead of a
spinner or a grid skeleton that swaps to an empty state.

Traces was previously documented as un-prefetchable because the project
scope is client state; createServerApiFactory already reads the same
active-project cookie, so the server can scope the fetch identically.
The wrapper now decides the empty state on first render from the
prefetched total instead of after the grid mounts.

Removes the isAuthenticated(status) -> setLoading(false) trap that
flashed an empty layout while the session was still resolving.
