+++ apps/frontend/src/app/(protected)/tokens/page.tsx
@@ -1,12 +1,21 @@
 import { Metadata } from 'next';
 import { Alert, Paper } from '@mui/material';
 import { auth } from '@/auth';
+import { createServerApiFactory } from '@/utils/api-client/server-factory';
+import { prefetchList } from '@/utils/server-prefetch';
+import { emptyFilters, listParams } from '@/utils/list';
 import TokensPageClient from './components/TokensPageClient';
+import { tokensList } from './components/list';
 
 export const metadata: Metadata = {
   title: 'API Tokens',
 };
 
+/**
+ * Server component: fetches the first page of tokens before rendering so the
+ * page arrives with content already in place -- no client-side spinner on
+ * first load. See `prefetchList` for the permission-gating rationale.
+ */
 export default async function TokensPage() {
   const session = await auth();
 
@@ -20,5 +29,25 @@ export default async function TokensPage() {
     );
   }
 
-  return <TokensPageClient />;
+  const factory = await createServerApiFactory();
+  const { initialData, initialTotalCount } = await prefetchList(
+    tokensList.capability,
+    () =>
+      tokensList.list(
+        factory,
+        listParams(tokensList, {
+          page: 1,
+          pageSize: tokensList.defaultPageSize,
+          sort: tokensList.defaultSort,
+          filters: emptyFilters(tokensList),
+        })
+      )
+  );
+
+  return (
+    <TokensPageClient
+      initialData={initialData}
+      initialTotalCount={initialTotalCount}
+    />
+  );
 }
