+++ apps/frontend/src/app/(protected)/traces/components/TracesClientWrapper.tsx
@@ -12,6 +12,7 @@ import TimelineOutlinedIcon from '@mui/icons-material/TimelineOutlined';
 import { useDocumentTitle } from '@/hooks/useDocumentTitle';
 import { BORDER_RADIUS, ELEVATION } from '@/styles/theme';
 import TracesClient from './TracesClient';
+import type { TraceSummary } from '@/utils/api-client/interfaces/telemetry';
 import AccessDenied from '@/components/common/AccessDenied';
 import PageLoadingState from '@/components/common/PageLoadingState';
 import { useCanWithStatus } from '@/components/common/Can';
@@ -22,12 +23,17 @@ interface TracesClientWrapperProps {
   currentUserId?: string;
   currentUserName?: string;
   currentUserPicture?: string;
+  /** Server-fetched first page -- when present, skips the initial client fetch. */
+  initialData?: TraceSummary[];
+  initialTotalCount?: number;
 }
 
 export default function TracesClientWrapper({
   currentUserId = '',
   currentUserName = '',
   currentUserPicture,
+  initialData,
+  initialTotalCount = 0,
 }: TracesClientWrapperProps) {
   const { status } = useSession();
   const searchParams = useSearchParams();
@@ -36,7 +42,12 @@ export default function TracesClientWrapper({
   const { allowed: canRead, loading: permsLoading } = useCanWithStatus(
     Capability.Telemetry.READ
   );
-  const [showEmptyHint, setShowEmptyHint] = useState(false);
+  // Decided on first render when the server already fetched page 1, so an
+  // empty project shows the empty state straight away instead of a grid
+  // skeleton that swaps out once the client fetch lands.
+  const [showEmptyHint, setShowEmptyHint] = useState(
+    initialData !== undefined && initialTotalCount === 0 && !initialProjectId
+  );
   const [refreshTrigger, setRefreshTrigger] = useState(0);
 
   useDocumentTitle('Traces');
@@ -124,6 +135,8 @@ export default function TracesClientWrapper({
               initialProjectId={initialProjectId}
               onUnfilteredEmpty={handleUnfilteredEmpty}
               refreshTrigger={refreshTrigger}
+              initialData={initialData}
+              initialTotalCount={initialTotalCount}
             />
           </Paper>
         </Box>
